rawurlencode/rawurlencode fixes #51
Paypal requires exactly the same raw post headers, and they use raw
encoding/decoding (always %20, not +). This means that urldecode in the
example gets time zones wrong (the + in +0100 for e.g. gets treated as a
space), and it re-encodes incorrectly with +'s where there were %20's.
(Why on earth isn't this done properly at the Paypal end to look at the
individual variables rather than the raw data!)